feat(app_client): add gRPC parameter validation and optimize log output
- Add validateGrpcParameters() function to check gRPC parameters
- Optimize log output for RPC options and channel arguments- Update Dijkstra algorithm to support custom distance function and float weights
- Refactor GrpcOptions class to improve code organization
- Clean up generated gRPC code

+++ common/filesystem/io/interface/IReadable.hpp
@@ -1,14 +1,19 @@
 #pragma once
+
+#include <cstdint>
+
 namespace fox {
 /// @brief Interface for readable objects.
 /// This interface defines the contract for objects that can be read from.
 /// Implementing classes should provide functionality to read data, typically byte by byte.
 class IReadable {
  public:
+  /// @brief Virtual destructor to ensure proper cleanup of derived classes
   virtual ~IReadable() = default;
 
-  /// @brief Read one byte
-  /// @return Byte value
-  virtual auto read() -> int32_t = 0;
+  /// @brief Read one byte from the input source.
+  /// This method reads and returns the next byte of data from the input source.
+  /// @return The byte value read, or -1 if the end of the stream has been reached
+  virtual auto read() -> size_t = 0;
 };
 }  // namespace fox
